Added Controller Abstract and fixed error in meeting
